in functional tests, check git config values
