+++ tests/functional/test_clone_token.py
@@ -4,18 +4,9 @@ import subprocess
 
 import pytest
 
+import helpers
+
 tool_args = ["--color", "never", "--ssh", "never"]
-
-
-def get_branch(dir: str):
-    result = subprocess.run(
-        ["git", "branch", "--show-current"],
-        cwd=dir,
-        capture_output=True,
-        text=True,
-        check=True,
-    )
-    return result.stdout.strip()
 
 
 @pytest.fixture()
@@ -46,4 +37,7 @@ Cloned repo:
 {expected_dir}
 """
     assert result.stdout == expected_stdout
-    assert get_branch(expected_dir) == "my-feature"
+    assert helpers.get_branch(expected_dir) == "my-feature"
+    assert helpers.get_config(expected_dir, "gimmegit.branch") == "my-feature"
+    assert helpers.get_config(expected_dir, "gimmegit.baseRemote") == "upstream"
+    assert helpers.get_config(expected_dir, "gimmegit.baseBranch") == "main"
